+++ apps/web/src/components/PoolCandidatesTable/JobPlacementDialog.tsx
@@ -13,7 +13,6 @@ import {
 import {
   FragmentType,
   PlacementType,
-  PoolCandidateStatus,
   getFragment,
   graphql,
 } from "@gc-digital-talent/graphql";
@@ -31,13 +30,7 @@ import { ROLE_NAME, useAuthorization } from "@gc-digital-talent/auth";
 import { isNotPlacedStatus, isQualifiedStatus } from "~/utils/poolCandidate";
 import poolCandidateMessages from "~/messages/poolCandidateMessages";
 import { checkRole } from "~/utils/teamUtils";
-
-export const PLACEMENT_TYPE_STATUSES = [
-  PoolCandidateStatus.PlacedCasual,
-  PoolCandidateStatus.PlacedIndeterminate,
-  PoolCandidateStatus.PlacedTentative,
-  PoolCandidateStatus.PlacedTerm,
-];
+import { PLACEMENT_TYPE_STATUSES } from "~/constants/poolCandidate";
 
 const PlaceCandidate_Mutation = graphql(/* GraphQL */ `
   mutation PlaceCandidate_Mutation(
